requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
imutils==0.5.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pycparser==2.20
pymongo==3.11.3
pytesseract==0.3.7
pyzbar==0.1.8
requests==2.25.1
urllib3==1.26.4
Werkzeug==1.0.1
zbar-py==1.0.4
